IntegratedCosetMds (#109)
Like `CosetMds`, with a few differences:
- (Bit reversed, a la Bowers) DIF + DIT rather than DIT + DIF
- We skip bit reversals of the inputs and outputs
- We don't weight by `1/N`, since this doesn't affect the MDS property
- We integrate the coset shifts into the DIF's twiddle factors

Also change `CosetMds` to precompute twiddles.

+++ mersenne-31/src/lib.rs
@@ -158,7 +158,7 @@ impl Field for Mersenne31 {
 
     fn mul_2exp_u64(&self, exp: u64) -> Self {
         // In a Mersenne field, multiplication by 2^k is just a left rotation by k bits.
-        let exp = (exp % 31) as u8;
+        let exp = exp % 31;
         let left = (self.value << exp) & ((1 << 31) - 1);
         let right = self.value >> (31 - exp);
         let rotated = left | right;
